moving tests around and merge conflicts
