+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/suites/hip869/UpdateAccountEnabledTest.java
@@ -5,12 +5,8 @@ import static com.hedera.services.bdd.junit.ContextRequirement.THROTTLE_OVERRIDE
 import static com.hedera.services.bdd.junit.EmbeddedReason.MUST_SKIP_INGEST;
 import static com.hedera.services.bdd.junit.EmbeddedReason.NEEDS_STATE_ACCESS;
 import static com.hedera.services.bdd.junit.TestTags.MATS;
-import static com.hedera.services.bdd.junit.TestTags.ONLY_SUBPROCESS;
-import static com.hedera.services.bdd.junit.hedera.utils.AddressBookUtils.CLASSIC_FIRST_NODE_ACCOUNT_NUM;
-import static com.hedera.services.bdd.spec.HapiPropertySource.asAccountString;
 import static com.hedera.services.bdd.spec.HapiSpec.hapiTest;
 import static com.hedera.services.bdd.spec.keys.TrieSigMapGenerator.uniqueWithFullPrefixesFor;
-import static com.hedera.services.bdd.spec.queries.QueryVerbs.getAccountInfo;
 import static com.hedera.services.bdd.spec.queries.QueryVerbs.getTxnRecord;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.cryptoCreate;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.nodeCreate;
@@ -24,9 +20,7 @@ import static com.hedera.services.bdd.suites.hip869.NodeCreateTest.generateX509C
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.INVALID_NODE_ACCOUNT_ID;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.INVALID_SIGNATURE;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import com.hedera.services.bdd.junit.EmbeddedHapiTest;
 import com.hedera.services.bdd.junit.HapiTest;
@@ -35,7 +29,6 @@ import com.hedera.services.bdd.junit.LeakyEmbeddedHapiTest;
 import com.hedera.services.bdd.junit.support.TestLifecycle;
 import com.hederahashgraph.api.proto.java.AccountID;
 import edu.umd.cs.findbugs.annotations.NonNull;
-import java.nio.file.Paths;
 import java.security.cert.CertificateEncodingException;
 import java.security.cert.X509Certificate;
 import java.util.List;
@@ -44,7 +37,6 @@ import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DynamicTest;
-import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Tag;
 
 /**
@@ -164,278 +156,5 @@ public class UpdateAccountEnabledTest {
                     assertEquals(
                             node.accountId().accountNum(), newAccountId.get().getAccountNum());
                 })));
-    }
-
-    @EmbeddedHapiTest(NEEDS_STATE_ACCESS)
-    final Stream<DynamicTest> updateAccountIdRequiredSignatures() {
-        final AtomicReference<AccountID> initialNodeAccountId = new AtomicReference<>();
-        final AtomicReference<AccountID> newAccountId = new AtomicReference<>();
-        return hapiTest(
-                newKeyNamed("adminKey"),
-                cryptoCreate("initialNodeAccount").exposingCreatedIdTo(initialNodeAccountId::set),
-                cryptoCreate("newAccount").exposingCreatedIdTo(newAccountId::set),
-                sourcing(() -> {
-                    try {
-                        return nodeCreate("testNode", "initialNodeAccount")
-                                .adminKey("adminKey")
-                                .gossipCaCertificate(
-                                        gossipCertificates.getFirst().getEncoded());
-                    } catch (CertificateEncodingException e) {
-                        throw new RuntimeException(e);
-                    }
-                }),
-                // signed with correct sig fails if account is sentinel
-                nodeUpdate("testNode")
-                        .accountId("0.0.0")
-                        .payingWith(DEFAULT_PAYER)
-                        .signedByPayerAnd("initialNodeAccount")
-                        .hasPrecheck(INVALID_NODE_ACCOUNT_ID),
-                // signed with correct sig passes if account is valid
-                nodeUpdate("testNode")
-                        .accountId("newAccount")
-                        .payingWith(DEFAULT_PAYER)
-                        .signedByPayerAnd("adminKey", "newAccount"),
-                viewNode("testNode", node -> assertEquals(toPbj(newAccountId.get()), node.accountId())),
-                // signed without adminKey works if only updating accountId
-                nodeUpdate("testNode")
-                        .accountId("initialNodeAccount")
-                        .payingWith(DEFAULT_PAYER)
-                        .signedByPayerAnd("newAccount", "initialNodeAccount"),
-                viewNode("testNode", node -> assertEquals(toPbj(initialNodeAccountId.get()), node.accountId())),
-                // signed without adminKey fails if updating other fields too
-                nodeUpdate("testNode")
-                        .accountId("newAccount")
-                        .description("updatedNode")
-                        .payingWith(DEFAULT_PAYER)
-                        .signedByPayerAnd("initialNodeAccount", "newAccount")
-                        .hasPrecheck(INVALID_SIGNATURE),
-                viewNode("testNode", node -> assertEquals(toPbj(initialNodeAccountId.get()), node.accountId())));
-    }
-
-    @EmbeddedHapiTest(NEEDS_STATE_ACCESS)
-    final Stream<DynamicTest> updateAccountIdIsIdempotent() {
-        final AtomicReference<AccountID> initialNodeAccountId = new AtomicReference<>();
-        final AtomicReference<AccountID> newAccountId = new AtomicReference<>();
-        return hapiTest(
-                newKeyNamed("adminKey"),
-                cryptoCreate("initialNodeAccount").exposingCreatedIdTo(initialNodeAccountId::set),
-                cryptoCreate("newAccount").exposingCreatedIdTo(newAccountId::set),
-                sourcing(() -> {
-                    try {
-                        return nodeCreate("testNode", "initialNodeAccount")
-                                .adminKey("adminKey")
-                                .gossipCaCertificate(
-                                        gossipCertificates.getFirst().getEncoded());
-                    } catch (CertificateEncodingException e) {
-                        throw new RuntimeException(e);
-                    }
-                }),
-                nodeUpdate("testNode")
-                        .accountId("newAccount")
-                        .payingWith(DEFAULT_PAYER)
-                        .signedByPayerAnd("adminKey", "newAccount"),
-                viewNode("testNode", node -> assertEquals(toPbj(newAccountId.get()), node.accountId())),
-                // node update with the same accountId should pass
-                nodeUpdate("testNode")
-                        .accountId("newAccount")
-                        .payingWith(DEFAULT_PAYER)
-                        .signedByPayerAnd("adminKey", "newAccount"),
-                viewNode("testNode", node -> assertEquals(toPbj(newAccountId.get()), node.accountId())));
-    }
-
-    @HapiTest
-    final Stream<DynamicTest> restrictNodeAccountDeletion() throws CertificateEncodingException {
-        final var adminKey = "adminKey";
-        final var account = "account";
-        final var secondAccount = "secondAccount";
-        final var node = "testNode";
-        return hapiTest(
-                cryptoCreate(account),
-                cryptoCreate(secondAccount),
-                newKeyNamed(adminKey),
-
-                // create new node
-                nodeCreate(node, account)
-                        .adminKey("adminKey")
-                        .gossipCaCertificate(gossipCertificates.getFirst().getEncoded()),
-                // verify we can't delete the node account
-                cryptoDelete(account).hasKnownStatus(ACCOUNT_IS_LINKED_TO_A_NODE),
-
-                // update the new node account id
-                nodeUpdate(node)
-                        .accountId(secondAccount)
-                        .payingWith(secondAccount)
-                        .signedBy(secondAccount, adminKey),
-
-                // verify now we can delete the old node account, and can't delete the new node account
-                cryptoDelete(account),
-                cryptoDelete(secondAccount).hasKnownStatus(ACCOUNT_IS_LINKED_TO_A_NODE),
-
-                // delete the node
-                nodeDelete(node).signedByPayerAnd(adminKey),
-                // verify we can delete the second account
-                cryptoDelete(secondAccount));
-    }
-
-    @HapiTest
-    final Stream<DynamicTest> nodeUpdateWithAccountLinkedToAnotherAccount() throws CertificateEncodingException {
-        final var adminKey = "adminKey";
-        final var account = "account";
-        final var secondAccount = "secondAccount";
-        final var node1 = "Node1";
-        final var node2 = "Node2";
-        return hapiTest(
-                cryptoCreate(account),
-                cryptoCreate(secondAccount),
-                newKeyNamed(adminKey),
-                nodeCreate(node1, account)
-                        .adminKey("adminKey")
-                        .gossipCaCertificate(gossipCertificates.getFirst().getEncoded()),
-                nodeCreate(node2, secondAccount)
-                        .adminKey("adminKey")
-                        .gossipCaCertificate(gossipCertificates.getFirst().getEncoded()),
-                // Verify node 1 update with second account will fail
-                nodeUpdate(node1)
-                        .accountId(secondAccount)
-                        .signedByPayerAnd(secondAccount, adminKey)
-                        .hasKnownStatus(ACCOUNT_IS_LINKED_TO_A_NODE),
-                // clear nodes from state
-                nodeDelete(node1).signedByPayerAnd(adminKey),
-                nodeDelete(node2).signedByPayerAnd(adminKey));
-    }
-
-    @HapiTest
-    final Stream<DynamicTest> nodeUpdateWithZeroBalanceAccount() throws CertificateEncodingException {
-        final var adminKey = "adminKey";
-        final var account = "account";
-        final var zeroBalanceAccount = "zeroBalanceAccount";
-        final var node = "testNode";
-        return hapiTest(
-                cryptoCreate(account),
-                cryptoCreate(zeroBalanceAccount).balance(0L),
-                newKeyNamed(adminKey),
-                nodeCreate(node, account)
-                        .adminKey("adminKey")
-                        .gossipCaCertificate(gossipCertificates.getFirst().getEncoded()),
-                // Verify node update with zero balance account will fail
-                nodeUpdate(node)
-                        .accountId(zeroBalanceAccount)
-                        .signedByPayerAnd(zeroBalanceAccount, adminKey)
-                        .hasKnownStatus(NODE_ACCOUNT_HAS_ZERO_BALANCE),
-                // Fund the account and try again
-                cryptoTransfer(movingHbar(ONE_HBAR).between(GENESIS, zeroBalanceAccount)),
-                nodeUpdate(node).accountId(zeroBalanceAccount).signedByPayerAnd(zeroBalanceAccount, adminKey),
-                // clear nodes from state
-                nodeDelete(node).signedByPayerAnd(adminKey));
-    }
-
-    @HapiTest
-    final Stream<DynamicTest> nodeUpdateWithDeletedAccount() throws CertificateEncodingException {
-        final var adminKey = "adminKey";
-        final var account = "account";
-        final var deletedAccount = "deletedAccount";
-        final var node = "testNode";
-        return hapiTest(
-                cryptoCreate(account),
-                newKeyNamed(adminKey),
-                nodeCreate(node, account)
-                        .adminKey("adminKey")
-                        .gossipCaCertificate(gossipCertificates.getFirst().getEncoded()),
-                cryptoCreate(deletedAccount),
-                cryptoDelete(deletedAccount),
-                // Verify node update will fail
-                nodeUpdate(node)
-                        .accountId(deletedAccount)
-                        .signedByPayerAnd(deletedAccount, adminKey)
-                        .hasKnownStatus(ACCOUNT_DELETED));
-    }
-
-    @Tag(ONLY_SUBPROCESS)
-    @HapiTest
-    final Stream<DynamicTest> accountUpdateBuildsProperRecordPath() {
-        final AtomicReference<AccountID> newAccountId = new AtomicReference<>();
-        final AtomicReference<AccountID> oldNodeAccountId = new AtomicReference<>();
-        final String nodeToUpdate = "3";
-        final String baseDir = "build/hapi-test/node" + nodeToUpdate + "/data/recordStreams/";
-
-        return hapiTest(
-                cryptoCreate("newAccount").exposingCreatedIdTo(newAccountId::set),
-                // account 6 is the node account of node 3
-                getAccountInfo("6").exposingIdTo(oldNodeAccountId::set),
-                nodeUpdate(nodeToUpdate).accountId("newAccount").signedByPayerAnd("newAccount"),
-                // create a transaction after the update so record files are generated
-                cryptoCreate("foo"),
-                // assert record paths
-                withOpContext((spec, log) -> {
-                    final var oldRecordPath = Paths.get(baseDir + "record" + asAccountString(oldNodeAccountId.get()));
-                    final var newRecordPath = Paths.get(baseDir + "record" + asAccountString(newAccountId.get()));
-                    assertTrue(oldRecordPath.toFile().exists());
-                    assertFalse(newRecordPath.toFile().exists());
-                }));
-    }
-
-    @HapiTest
-    final Stream<DynamicTest> updateAccountToAdmin() throws CertificateEncodingException {
-        final var adminKey = "adminKey";
-        return hapiTest(
-                newKeyNamed(adminKey),
-                cryptoCreate("nodeAccount"),
-                cryptoCreate("newAccount").key(adminKey),
-                nodeCreate("testNode", "nodeAccount")
-                        .adminKey(adminKey)
-                        .gossipCaCertificate(gossipCertificates.getFirst().getEncoded()),
-                nodeUpdate("testNode").accountId("newAccount").signedByPayerAnd(adminKey));
-    }
-
-    @Nested
-    public class NegativeCases {
-
-        @HapiTest
-        final Stream<DynamicTest> updateAccountNegSignatures() throws CertificateEncodingException {
-            return hapiTest(
-                    newKeyNamed("adminKey"),
-                    cryptoCreate("nodeAccount"),
-                    cryptoCreate("newAccount"),
-                    nodeCreate("testNode", "nodeAccount")
-                            .adminKey("adminKey")
-                            .gossipCaCertificate(gossipCertificates.getFirst().getEncoded()),
-
-                    // not signed by 'newAccount' so it should fail
-                    nodeUpdate("testNode")
-                            .accountId("newAccount")
-                            .signedByPayerAnd("adminKey")
-                            .hasKnownStatus(INVALID_SIGNATURE),
-                    // not signed by 'adminKey' so it should fail
-                    nodeUpdate("testNode")
-                            .accountId("newAccount")
-                            .signedByPayerAnd("newAccount")
-                            .hasKnownStatus(INVALID_SIGNATURE),
-                    // signed with both passes
-                    nodeUpdate("testNode").accountId("newAccount").signedByPayerAnd("adminKey", "newAccount"));
-        }
-
-        @HapiTest
-        final Stream<DynamicTest> updateAccountNegSignaturesWithoutAdminKey() throws CertificateEncodingException {
-            return hapiTest(
-                    newKeyNamed("adminKey"),
-                    cryptoCreate("nodeAccount"),
-                    cryptoCreate("newAccount"),
-                    nodeCreate("testNode", "nodeAccount")
-                            .adminKey("adminKey")
-                            .gossipCaCertificate(gossipCertificates.getFirst().getEncoded()),
-
-                    // not signed by 'newAccount' so it should fail
-                    nodeUpdate("testNode")
-                            .accountId("newAccount")
-                            .signedByPayerAnd("nodeAccount")
-                            .hasKnownStatus(INVALID_SIGNATURE),
-                    // not signed by 'nodeAccount' so it should fail
-                    nodeUpdate("testNode")
-                            .accountId("newAccount")
-                            .signedByPayerAnd("newAccount")
-                            .hasKnownStatus(INVALID_SIGNATURE),
-                    // signed with both passes
-                    nodeUpdate("testNode").accountId("newAccount").signedByPayerAnd("nodeAccount", "newAccount"));
-        }
     }
 }
